Reimplement metadata, field and fieldlist for netcdf
