+++ tests/grib/test_grib_metadata.py
@@ -15,9 +15,7 @@ import numpy as np
 import pytest
 
 from earthkit.data import from_source
-from earthkit.data.core.temporary import temp_file
 from earthkit.data.testing import earthkit_examples_file, earthkit_test_data_file
-from earthkit.data.utils import projections
 
 
 def check_array(v, shape=None, first=None, last=None, meanv=None, eps=1e-3):
@@ -485,344 +483,6 @@ def test_grib_metadata_namespace():
     assert "must be a str when key specified" in str(excinfo.value)
 
 
-def test_grib_values_1():
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    eps = 1e-5
-
-    # whole file
-    v = f.values
-    assert isinstance(v, np.ndarray)
-    assert v.shape == (1, 84)
-    v = v[0].flatten()
-    check_array(
-        v,
-        (84,),
-        first=260.43560791015625,
-        last=227.18560791015625,
-        meanv=274.36566743396577,
-        eps=eps,
-    )
-
-    # field
-    v1 = f[0].values
-    assert isinstance(v1, np.ndarray)
-    assert v1.shape == (84,)
-    assert np.allclose(v, v1, eps)
-
-
-def test_grib_values_18():
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-
-    eps = 1e-5
-
-    # whole file
-    v = f.values
-    assert isinstance(v, np.ndarray)
-    assert v.shape == (18, 84)
-    vf = v[0].flatten()
-    check_array(
-        vf,
-        (84,),
-        first=272.5642,
-        last=240.56417846679688,
-        meanv=279.70703560965404,
-        eps=eps,
-    )
-
-    vf = v[15].flatten()
-    check_array(
-        vf,
-        (84,),
-        first=226.6531524658203,
-        last=206.6531524658203,
-        meanv=227.84362865629652,
-        eps=eps,
-    )
-
-
-def test_grib_to_numpy_1():
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    eps = 1e-5
-    v = f.to_numpy()
-    assert isinstance(v, np.ndarray)
-    v = v[0].flatten()
-    check_array(
-        v,
-        (84,),
-        first=260.43560791015625,
-        last=227.18560791015625,
-        meanv=274.36566743396577,
-        eps=eps,
-    )
-
-
-@pytest.mark.parametrize(
-    "first,options, expected_shape",
-    [
-        (False, {}, (1, 7, 12)),
-        (False, {"flatten": True}, (1, 84)),
-        (False, {"flatten": False}, (1, 7, 12)),
-        (True, {}, (7, 12)),
-        (True, {"flatten": True}, (84,)),
-        (True, {"flatten": False}, (7, 12)),
-    ],
-)
-def test_grib_to_numpy_1_shape(first, options, expected_shape):
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    v_ref = f[0].to_numpy().flatten()
-    eps = 1e-5
-
-    data = f[0] if first else f
-    v1 = data.to_numpy(**options)
-    assert isinstance(v1, np.ndarray)
-    assert v1.shape == expected_shape
-    v1 = v1.flatten()
-    assert np.allclose(v_ref, v1, eps)
-
-
-def test_grib_to_numpy_18():
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-
-    eps = 1e-5
-
-    # whole file
-    v = f.to_numpy(flatten=True)
-    assert isinstance(v, np.ndarray)
-    assert v.shape == (18, 84)
-    vf0 = v[0].flatten()
-    check_array(
-        vf0,
-        (84,),
-        first=272.5642,
-        last=240.56417846679688,
-        meanv=279.70703560965404,
-        eps=eps,
-    )
-
-    vf15 = v[15].flatten()
-    check_array(
-        vf15,
-        (84,),
-        first=226.6531524658203,
-        last=206.6531524658203,
-        meanv=227.84362865629652,
-        eps=eps,
-    )
-
-
-@pytest.mark.parametrize(
-    "options, expected_shape",
-    [
-        (
-            {},
-            (
-                18,
-                7,
-                12,
-            ),
-        ),
-        (
-            {"flatten": True},
-            (
-                18,
-                84,
-            ),
-        ),
-        ({"flatten": False}, (18, 7, 12)),
-    ],
-)
-def test_grib_to_numpy_18_shape(options, expected_shape):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-
-    eps = 1e-5
-
-    # whole file
-    v = f.to_numpy()
-    assert isinstance(v, np.ndarray)
-    assert v.shape == (18, 7, 12)
-    vf0 = f[0].to_numpy().flatten()
-    assert vf0.shape == (84,)
-    vf15 = f[15].to_numpy().flatten()
-    assert vf15.shape == (84,)
-
-    v1 = f.to_numpy(**options)
-    assert isinstance(v1, np.ndarray)
-    assert v1.shape == expected_shape
-    vr = v1[0].flatten()
-    assert np.allclose(vf0, vr, eps)
-    vr = v1[15].flatten()
-    assert np.allclose(vf15, vr, eps)
-
-
-@pytest.mark.parametrize("dtype", [np.float32, np.float64])
-def test_grib_to_numpy_1_dtype(dtype):
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    v = f[0].to_numpy(dtype=dtype)
-    assert v.dtype == dtype
-
-    v = f.to_numpy(dtype=dtype)
-    assert v.dtype == dtype
-
-
-@pytest.mark.parametrize("dtype", [np.float32, np.float64])
-def test_grib_to_numpy_18_dtype(dtype):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-
-    v = f[0].to_numpy(dtype=dtype)
-    assert v.dtype == dtype
-
-    v = f.to_numpy(dtype=dtype)
-    assert v.dtype == dtype
-
-
-def test_grib_values_with_missing():
-    f = from_source("file", earthkit_test_data_file("test_single_with_missing.grib"))
-
-    v = f[0].values
-    assert isinstance(v, np.ndarray)
-    assert v.shape == (84,)
-    eps = 0.001
-    assert np.count_nonzero(np.isnan(v)) == 38
-    mask = np.array([12, 14, 15, 24, 25, 26] + list(range(28, 60)))
-    assert np.isclose(v[0], 260.4356, eps)
-    assert np.isclose(v[11], 260.4356, eps)
-    assert np.isclose(v[-1], 227.1856, eps)
-    m = v[mask]
-    assert len(m) == 38
-    assert np.count_nonzero(np.isnan(m)) == 38
-
-
-@pytest.mark.parametrize("index", [0, None])
-def test_grib_to_latlon_single(index):
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    eps = 1e-5
-    g = f[index] if index is not None else f
-    v = g.to_latlon(flatten=True)
-    assert isinstance(v, dict)
-    assert isinstance(v["lon"], np.ndarray)
-    assert isinstance(v["lat"], np.ndarray)
-    check_array(
-        v["lon"],
-        (84,),
-        first=0.0,
-        last=330.0,
-        meanv=165.0,
-        eps=eps,
-    )
-    check_array(
-        v["lat"],
-        (84,),
-        first=90,
-        last=-90,
-        meanv=0,
-        eps=eps,
-    )
-
-
-@pytest.mark.parametrize("index", [0, None])
-def test_grib_to_latlon_single_shape(index):
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    g = f[index] if index is not None else f
-    v = g.to_latlon()
-    assert isinstance(v, dict)
-    assert isinstance(v["lon"], np.ndarray)
-    assert isinstance(v["lat"], np.ndarray)
-
-    # x
-    assert v["lon"].shape == (7, 12)
-    for x in v["lon"]:
-        assert np.allclose(x, np.linspace(0, 330, 12))
-
-    # y
-    assert v["lat"].shape == (7, 12)
-    for i, y in enumerate(v["lat"]):
-        assert np.allclose(y, np.ones(12) * (90 - i * 30))
-
-
-def test_grib_to_latlon_multi():
-    f = from_source("file", earthkit_examples_file("test.grib"))
-
-    v_ref = f[0].to_latlon(flatten=True)
-    v = f.to_latlon(flatten=True)
-    assert isinstance(v, dict)
-    assert v.keys() == v_ref.keys()
-
-    assert isinstance(v, dict)
-    assert np.allclose(v["lat"], v_ref["lat"])
-    assert np.allclose(v["lon"], v_ref["lon"])
-
-
-def test_grib_to_latlon_multi_non_shared_grid():
-    f1 = from_source("file", earthkit_examples_file("test.grib"))
-    f2 = from_source("file", earthkit_examples_file("test4.grib"))
-    f = f1 + f2
-
-    with pytest.raises(ValueError):
-        f.to_latlon()
-
-
-@pytest.mark.parametrize("index", [0, None])
-def test_grib_to_points_single(index):
-    f = from_source("file", earthkit_test_data_file("test_single.grib"))
-
-    eps = 1e-5
-    g = f[index] if index is not None else f
-    v = g.to_points(flatten=True)
-    assert isinstance(v, dict)
-    assert isinstance(v["x"], np.ndarray)
-    assert isinstance(v["y"], np.ndarray)
-    check_array(
-        v["x"],
-        (84,),
-        first=0.0,
-        last=330.0,
-        meanv=165.0,
-        eps=eps,
-    )
-    check_array(
-        v["y"],
-        (84,),
-        first=90,
-        last=-90,
-        meanv=0,
-        eps=eps,
-    )
-
-
-def test_grib_to_points_unsupported_grid():
-    f = from_source("file", earthkit_test_data_file("mercator.grib"))
-    with pytest.raises(ValueError):
-        f[0].to_points()
-
-
-def test_grib_to_points_multi():
-    f = from_source("file", earthkit_examples_file("test.grib"))
-
-    v_ref = f[0].to_points(flatten=True)
-    v = f.to_points(flatten=True)
-    assert isinstance(v, dict)
-    assert v.keys() == v_ref.keys()
-
-    assert isinstance(v, dict)
-    assert np.allclose(v["x"], v_ref["x"])
-    assert np.allclose(v["y"], v_ref["y"])
-
-
-def test_grib_to_points_multi_non_shared_grid():
-    f1 = from_source("file", earthkit_examples_file("test.grib"))
-    f2 = from_source("file", earthkit_examples_file("test4.grib"))
-    f = f1 + f2
-
-    with pytest.raises(ValueError):
-        f.to_points()
-
-
 def test_grib_datetime():
     s = from_source("file", earthkit_examples_file("test.grib"))
 
@@ -852,41 +512,6 @@ def test_grib_datetime():
     assert s.datetime() == ref
 
 
-def test_bbox():
-    ds = from_source("file", earthkit_examples_file("test.grib"))
-    bb = ds.bounding_box()
-    assert len(bb) == 2
-    for b in bb:
-        assert b.as_tuple() == (73, -27, 33, 45)
-
-
-@pytest.mark.parametrize("index", [0, None])
-def test_grib_projection_ll(index):
-    f = from_source("file", earthkit_examples_file("test.grib"))
-
-    if index is not None:
-        g = f[index]
-    else:
-        g = f
-    assert isinstance(
-        g.projection(), (projections.EquidistantCylindrical, projections.LongLat)
-    )
-
-
-def test_grib_projection_mercator():
-    f = from_source("file", earthkit_test_data_file("mercator.grib"))
-    projection = f[0].projection()
-    assert isinstance(projection, projections.Mercator)
-    assert projection.parameters == {
-        "true_scale_latitude": 20,
-        "central_latitude": 0,
-        "central_longitude": 0,
-        "false_easting": 0,
-        "false_northing": 0,
-    }
-    assert projection.globe == dict()
-
-
 def test_message():
     f = from_source("file", earthkit_examples_file("test.grib"))
     v = f[0].message()
@@ -895,146 +520,6 @@ def test_message():
     v = f[1].message()
     assert len(v) == 526
     assert v[:4] == b"GRIB"
-
-
-def test_grib_from_memory():
-    with open(earthkit_test_data_file("test_single.grib"), "rb") as f:
-        data = f.read()
-        fs = from_source("memory", data)
-        assert len(fs) == 1
-        sn = fs.metadata("param")
-        assert sn == ["2t"]
-
-
-def test_grib_from_stream_invalid_args():
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        with pytest.raises(TypeError):
-            from_source("stream", stream, order_by="level")
-
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        with pytest.raises(TypeError):
-            from_source("stream", stream, group_by=1)
-
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        with pytest.raises(TypeError):
-            from_source("stream", stream, group_by=["level", 1])
-
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        with pytest.raises(TypeError):
-            from_source("stream", stream, group_by="level", batch_size=1)
-
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        with pytest.raises(ValueError):
-            from_source("stream", stream, batch_size=-1)
-
-
-@pytest.mark.parametrize("group_by", ["level", ["level", "gridType"]])
-def test_grib_from_stream_group_by(group_by):
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        fs = from_source("stream", stream, group_by=group_by)
-
-        # no methods are available
-        with pytest.raises(TypeError):
-            len(fs)
-
-        ref = [
-            [("t", 1000), ("u", 1000), ("v", 1000)],
-            [("t", 850), ("u", 850), ("v", 850)],
-        ]
-        for i, f in enumerate(fs):
-            assert len(f) == 3
-            assert f.metadata(("param", "level")) == ref[i]
-
-        # stream consumed, no data is available
-        assert sum([1 for _ in fs]) == 0
-
-
-def test_grib_from_stream_single_batch():
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        fs = from_source("stream", stream)
-
-        # no methods are available
-        with pytest.raises(TypeError):
-            len(fs)
-
-        ref = ["t", "u", "v", "t", "u", "v"]
-        val = []
-        for f in fs:
-            v = f.metadata("param")
-            val.append(v)
-
-        assert val == ref
-
-        # stream consumed, no data is available
-        assert sum([1 for _ in fs]) == 0
-
-
-def test_grib_from_stream_multi_batch():
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        fs = from_source("stream", stream, batch_size=2)
-
-        # no methods are available
-        with pytest.raises(TypeError):
-            len(fs)
-
-        ref = [["t", "u"], ["v", "t"], ["u", "v"]]
-        for i, f in enumerate(fs):
-            assert len(f) == 2
-            f.metadata("param") == ref[i]
-
-        # stream consumed, no data is available
-        assert sum([1 for _ in fs]) == 0
-
-
-def test_grib_from_stream_in_memory():
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        fs = from_source("stream", stream, batch_size=0)
-
-        assert len(fs) == 6
-
-        ref = ["t", "u", "v", "t", "u", "v"]
-        val = []
-
-        # iteration
-        for f in fs:
-            v = f.metadata("param")
-            val.append(v)
-
-        assert val == ref, "iteration"
-
-        # metadata
-        val = []
-        val = fs.metadata("param")
-        assert val == ref, "method"
-
-
-def test_grib_save_when_loaded_from_file():
-    fs = from_source("file", earthkit_examples_file("test6.grib"))
-    assert len(fs) == 6
-    with temp_file() as tmp:
-        fs.save(tmp)
-        fs_saved = from_source("file", tmp)
-        assert len(fs) == len(fs_saved)
-
-
-def test_grib_save_when_loaded_from_memory():
-    with open(earthkit_test_data_file("test_single.grib"), "rb") as f:
-        data = f.read()
-        fs = from_source("memory", data)
-        with temp_file() as tmp:
-            fs.save(tmp)
-            fs_saved = from_source("file", tmp)
-            assert len(fs) == len(fs_saved)
-
-
-def test_grib_save_when_loaded_from_stream():
-    with open(earthkit_examples_file("test6.grib"), "rb") as stream:
-        fs = from_source("stream", stream, batch_size=0)
-        assert len(fs) == 6
-        with temp_file() as tmp:
-            fs.save(tmp)
-            fs_saved = from_source("file", tmp)
-            assert len(fs) == len(fs_saved)
 
 
 if __name__ == "__main__":
