+++ tests/netcdf/test_netcdf_slice.py
@@ -9,7 +9,6 @@
 # nor does it submit to any jurisdiction.
 #
 
-
 import numpy as np
 import pytest
 
@@ -21,25 +20,25 @@ from earthkit.data.testing import earthkit_examples_file
     "index,expected_meta",
     [
         (0, ["t", 1000]),
-        (2, ["v", 1000]),
+        (2, ["t", 700]),
         (17, ["v", 300]),
         (-1, ["v", 300]),
-        (-5, ["u", 400]),
+        (-5, ["v", 850]),
     ],
 )
-def test_grib_single_index(index, expected_meta):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+def test_netcdf_single_index(index, expected_meta):
+    f = from_source("file", earthkit_examples_file("tuv_pl.nc"))
 
     r = f[index]
-    assert r.metadata(["shortName", "level"]) == expected_meta
+    assert r.metadata(["variable", "level"]) == expected_meta
     v = r.values
     assert v.shape == (84,)
     # eps = 0.001
     # assert np.isclose(v[1088], 304.5642, eps)
 
 
-def test_grib_single_index_bad():
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+def test_netcdf_single_index_bad():
+    f = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     with pytest.raises(IndexError):
         f[27]
 
@@ -47,26 +46,27 @@ def test_grib_single_index_bad():
 @pytest.mark.parametrize(
     "indexes,expected_meta",
     [
-        (slice(0, 4), [["t", 1000], ["u", 1000], ["v", 1000], ["t", 850]]),
-        (slice(None, 4), [["t", 1000], ["u", 1000], ["v", 1000], ["t", 850]]),
-        (slice(2, 9, 2), [["v", 1000], ["u", 850], ["t", 700], ["v", 700]]),
-        (slice(8, 1, -2), [["v", 700], ["t", 700], ["u", 850], ["v", 1000]]),
-        (slice(14, 18), [["v", 400], ["t", 300], ["u", 300], ["v", 300]]),
-        (slice(14, None), [["v", 400], ["t", 300], ["u", 300], ["v", 300]]),
+        (slice(0, 4), [["t", 1000], ["t", 850], ["t", 700], ["t", 500]]),
+        (slice(None, 4), [["t", 1000], ["t", 850], ["t", 700], ["t", 500]]),
+        (slice(2, 9, 2), [["t", 700], ["t", 400], ["u", 1000], ["u", 700]]),
+        (slice(8, 1, -2), [["u", 700], ["u", 1000], ["t", 400], ["t", 700]]),
+        (slice(14, 18), [["v", 700], ["v", 500], ["v", 400], ["v", 300]]),
+        (slice(14, None), [["v", 700], ["v", 500], ["v", 400], ["v", 300]]),
     ],
 )
-def test_grib_slice_single_file(indexes, expected_meta):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+def test_netcdf_slice_single_file(indexes, expected_meta):
+    f = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     r = f[indexes]
     assert len(r) == 4
-    assert r.metadata(["shortName", "level"]) == expected_meta
+    assert r.metadata(["variable", "level"]) == expected_meta
     v = r.values
     assert v.shape == (4, 84)
     # check the original fieldlist
     assert len(f) == 18
-    assert f.metadata("shortName") == ["t", "u", "v"] * 6
+    assert f.metadata("variable") == ["t"] * 6 + ["u"] * 6 + ["v"] * 6
 
 
+@pytest.mark.xfail
 @pytest.mark.parametrize(
     "indexes,expected_meta",
     [
@@ -76,18 +76,18 @@ def test_grib_slice_single_file(indexes, expected_meta):
         (slice(3, 6), [["z", 500], ["t", 850], ["z", 850]]),
     ],
 )
-def test_grib_slice_multi_file(indexes, expected_meta):
+def test_netcdf_slice_multi_file(indexes, expected_meta):
     f = from_source(
         "file",
-        [earthkit_examples_file("test.grib"), earthkit_examples_file("test4.grib")],
+        [earthkit_examples_file("test.nc"), earthkit_examples_file("tuv_pl.nc")],
     )
     r = f[indexes]
     assert len(r) == 3
-    assert r.metadata(["shortName", "level"]) == expected_meta
+    assert r.metadata(["variable", "level"]) == expected_meta
     # v = r.values
     # assert v.shape == (3, 84)
     # check the original fieldlist
-    assert len(f) == 6
+    assert len(f) == 2 + 18
     assert f.metadata("shortName") == ["2t", "msl", "t", "z", "t", "z"]
 
 
@@ -95,40 +95,40 @@ def test_grib_slice_multi_file(indexes, expected_meta):
     "indexes1,indexes2",
     [(np.array([1, 16, 5, 9]), np.array([1, 3])), ([1, 16, 5, 9], [1, 3])],
 )
-def test_grib_array_indexing(indexes1, indexes2):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+def test_netcdf_array_indexing(indexes1, indexes2):
+    f = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     r = f[indexes1]
     assert len(r) == 4
-    assert r.metadata("shortName") == ["u", "u", "v", "t"]
+    assert r.metadata("variable") == ["t", "v", "t", "u"]
 
     r1 = r[indexes2]
     assert len(r1) == 2
-    assert r1.metadata("shortName") == ["u", "t"]
+    assert r1.metadata("variable") == ["v", "u"]
 
 
 @pytest.mark.parametrize("indexes", [(np.array([1, 19, 5, 9])), ([1, 19, 5, 9])])
-def test_grib_array_indexing_bad(indexes):
-    f = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+def test_netcdf_array_indexing_bad(indexes):
+    f = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     with pytest.raises(IndexError):
         f[indexes]
 
 
-def test_grib_fieldlist_iterator():
-    g = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-    sn = g.metadata("shortName")
+def test_netcdf_fieldlist_iterator():
+    g = from_source("file", earthkit_examples_file("tuv_pl.nc"))
+    sn = g.metadata("variable")
     assert len(sn) == 18
-    iter_sn = [f.metadata("shortName") for f in g]
+    iter_sn = [f.metadata("variable") for f in g]
     assert iter_sn == sn
     # repeated iteration
-    iter_sn = [f.metadata("shortName") for f in g]
+    iter_sn = [f.metadata("variable") for f in g]
     assert iter_sn == sn
 
 
-def test_grib_fieldlist_iterator_with_zip():
+def test_netcdf_fieldlist_iterator_with_zip():
     # this tests something different with the iterator - this does not try to
     # 'go off the edge' of the fieldlist, because the length is determined by
     # the list of levels
-    g = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+    g = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     ref_levs = g.metadata("level")
     assert len(ref_levs) == 18
     levs1 = []
@@ -140,9 +140,9 @@ def test_grib_fieldlist_iterator_with_zip():
     assert levs2 == ref_levs
 
 
-def test_grib_fieldlist_iterator_with_zip_multiple():
+def test_netcdf_fieldlist_iterator_with_zip_multiple():
     # same as test_fieldlist_iterator_with_zip() but multiple times
-    g = from_source("file", earthkit_examples_file("tuv_pl.grib"))
+    g = from_source("file", earthkit_examples_file("tuv_pl.nc"))
     ref_levs = g.metadata("level")
     assert len(ref_levs) == 18
     for i in range(2):
@@ -155,14 +155,20 @@ def test_grib_fieldlist_iterator_with_zip_multiple():
         assert levs2 == ref_levs, i
 
 
-def test_grib_fieldlist_reverse_iterator():
-    g = from_source("file", earthkit_examples_file("tuv_pl.grib"))
-    sn = g.metadata("shortName")
+def test_netcdf_fieldlist_reverse_iterator():
+    g = from_source("file", earthkit_examples_file("tuv_pl.nc"))
+    sn = g.metadata("variable")
     sn_reversed = list(reversed(sn))
     assert sn_reversed[0] == "v"
     assert sn_reversed[17] == "t"
     gr = reversed(g)
-    iter_sn = [f.metadata("shortName") for f in gr]
+    iter_sn = [f.metadata("variable") for f in gr]
     assert len(iter_sn) == len(sn_reversed)
     assert iter_sn == sn_reversed
-    assert iter_sn == ["v", "u", "t"] * 6
+    assert iter_sn == ["v"] * 6 + ["u"] * 6 + ["t"] * 6
+
+
+if __name__ == "__main__":
+    from earthkit.data.testing import main
+
+    main()
